Fixup

+++ gulp_tasks/client_post_build.ts
@@ -1,9 +1,11 @@
-import {CLIENT_POST_BUILD_CLEANUP} from './config';
+import {ENV, CLIENT_POST_BUILD_CLEANUP} from './config';
 
 export = function client_post_build(gulp, plugins, option) {
     return () => {
-        gulp
-            .src(CLIENT_POST_BUILD_CLEANUP, {read: false})
-            .pipe(plugins.rimraf());
+        if(ENV !== 'dev') {
+            gulp
+                .src(CLIENT_POST_BUILD_CLEANUP, {read: false})
+                .pipe(plugins.rimraf());
+        }
     };
 };
